+++ src/main/scala/org/mmisw/orr/ont/service/httpUtil.scala
@@ -1,5 +1,7 @@
 package org.mmisw.orr.ont.service
 
+import java.io.File
+
 import com.typesafe.scalalogging.{StrictLogging ⇒ Logging}
 
 import scala.util.control.NonFatal
@@ -7,8 +9,14 @@ import scalaj.http.{Http, HttpOptions, HttpResponse}
 
 object httpUtil extends AnyRef with Logging {
 
-  def downloadUrl(remoteUrl: String, acceptList: List[String] = List.empty)
-  : Either[Throwable,String] = {
+  case class DownloadResult(body:String,
+                            contentType: Option[String]
+                           )
+
+  def downloadUrl(remoteUrl: String,
+                  acceptList: List[String] = List.empty,
+                  saveInFile: Option[File] = None
+                 ): Either[Throwable, DownloadResult] = {
 
     logger.debug(s"downloadUrl: $remoteUrl  acceptList=$acceptList")
 
@@ -23,9 +31,13 @@ object httpUtil extends AnyRef with Logging {
       }
 
       val response: HttpResponse[String] = request.asString
-
-      if (response.code == 200)
-        Right(response.body)
+      if (response.code == 200) {
+        saveInFile foreach { file ⇒
+          java.nio.file.Files.write(file.toPath,
+            response.body.getBytes(java.nio.charset.StandardCharsets.UTF_8))
+        }
+        Right(DownloadResult(response.body, response.contentType))
+      }
       else {
         logger.debug(s"downloadUrl: $remoteUrl  status=${response.code}")
         val maxLen = 4000